requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
gunicorn==22.0.0

pydantic==2.9.2
pydantic-settings==2.6.1

google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-auth==2.35.0
requests==2.32.3

stripe==10.12.0
httpx==0.27.2

python-multipart==0.0.12

pytest==8.3.3

twilio==9.2.3
